Enable inlining on ARM64
The biggest thing that needed fixing up was the loading of inlinee
call info data; I implemented approximately the same thing that we
do on arm, minus doing it a little earlier (when encoding, instead
of in a relocation step). The most notable change is that I had to
make LabelInstr offsets grow to uintptr_t (from uint32); this will
not take any more memory, due to already being unioned with a byte
pointer, but has impact outside of ARM64.

The other part is the definition of two new instructions that have
a move with a known shift amount (movz_shift and movk_shift); this
is to ease definition of these fixed data moves, and may be useful
elsewhere.

+++ lib/Backend/IR.h
@@ -679,7 +679,7 @@ private:
     union labelLocation
     {
         BYTE *                  pc;     // Used by encoder and is the real pc offset
-        uint32                  offset; // Used by preEncoder and is an estimation pc offset, not accurate
+        uintptr_t               offset; // Used by preEncoder and is an estimation pc offset, not accurate
     } m_pc;
 
     BasicBlock *            m_block;
@@ -689,9 +689,9 @@ public:
 
     inline void             SetPC(BYTE * pc);
     inline BYTE *           GetPC(void) const;
-    inline void             SetOffset(uint32 offset);
-    inline void             ResetOffset(uint32 offset);
-    inline uint32           GetOffset(void) const;
+    inline void             SetOffset(uintptr_t offset);
+    inline void             ResetOffset(uintptr_t offset);
+    inline uintptr_t        GetOffset(void) const;
     inline void             SetBasicBlock(BasicBlock * block);
     inline BasicBlock *     GetBasicBlock(void) const;
     inline void             SetLoop(Loop *loop);
